Changed Folderstructrue. moved class files into new folder definitions.
