+++ src/CrossCutting/Logging/Logging/Data/EnumLogLevels.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MarkusMeinhard.Doci.CrossCutting.Logger
+namespace MarkusMeinhard.Doci.CrossCutting.Logging.Data
 {
 
     public enum LogLevels{
