+++ src/CrossCutting/Logging/Logging/Logging.csproj
@@ -41,13 +41,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ConsoleLogger.cs" />
-    <Compile Include="EnumLogLevels.cs" />
-    <Compile Include="ILogger.cs" />
-    <Compile Include="LogfilesController.cs" />
+    <Compile Include="Loggers\ConsoleLogger.cs" />
+    <Compile Include="Data\EnumLogLevels.cs" />
+    <Compile Include="Loggers\ILogger.cs" />
+    <Compile Include="Loggers\LogfilesController.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TextboxLogger.cs" />
-    <Compile Include="TextFileLogger.cs" />
+    <Compile Include="Loggers\TextboxLogger.cs" />
+    <Compile Include="Loggers\TextFileLogger.cs" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
